% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{compute_association_strength}
\alias{compute_association_strength}
\title{Compute Association Strength from a co-occurrence matrix}
\usage{
compute_association_strength(cooc_matrix, freq)
}
\arguments{
\item{cooc_matrix}{Symmetric sparse \code{Matrix} of co-occurrence counts as
returned by \code{\link{build_cooccurrence_matrix}}.}

\item{freq}{Named numeric (or integer) vector of raw term frequencies. Names
must match the row/column names of \code{cooc_matrix}.}
}
\value{
A sparse \code{dgCMatrix} of the same dimensions as
\code{cooc_matrix} containing AS values. Values are always in \eqn{[0,1]}.
}
\description{
Normalizes raw co-occurrence counts using the Association Strength measure:
\deqn{sA(i,j) = \frac{c(i,j)^2}{f(i) \cdot f(j)}}
where \eqn{c(i,j)} is the number of co-occurring sentences and \eqn{f(i)}
is the raw term frequency of term \eqn{i}.
}
\examples{
# Minimal example
m <- Matrix::sparseMatrix(
  i = c(1, 2), j = c(2, 1), x = c(3, 3),
  dims = c(3, 3),
  dimnames = list(c("a","b","c"), c("a","b","c"))
)
f <- c(a = 5L, b = 4L, c = 2L)
compute_association_strength(m, f)

}
