% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_coherence}
\alias{compute_coherence}
\title{Compute edge-density coherence for each community}
\usage{
compute_coherence(graph, communities)
}
\arguments{
\item{graph}{An \code{igraph} object.}

\item{communities}{Named list of character vectors (community members).}
}
\value{
Named numeric vector of coherence values in \eqn{[0, 1]}, one per
community. Communities with fewer than 2 members return \code{NA}.
}
\description{
Coherence is the ratio of actual edges to the maximum possible edges within
a community (i.e., graph edge density):
\deqn{\text{coherence}(G_i) = \frac{|E_i|}{|G_i|(|G_i|-1)/2}}
}
\examples{
\dontrun{
coh <- compute_coherence(graph, comm$communities)
}

}
