% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themescope.R
\name{print.themescope}
\alias{print.themescope}
\title{Print a themescope object}
\usage{
\method{print}{themescope}(x, ...)
}
\arguments{
\item{x}{A \code{themescope} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a concise summary of the ThemeScope analysis.
}
