% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themescope.R
\name{summary.themescope}
\alias{summary.themescope}
\title{Summarise a themescope object}
\usage{
\method{summary}{themescope}(object, ...)
}
\arguments{
\item{object}{A \code{themescope} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns a data frame of community statistics.
}
\description{
Provides a detailed summary including a table of community statistics.
}
