% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zscore}
\alias{zscore}
\title{Compute z-scores}
\usage{
zscore(x)
}
\arguments{
\item{x}{Numeric vector to standardize.}
}
\value{
Numeric vector of z-scores. NA values in input remain NA in output.
Returns a vector of NAs if standard deviation is zero.
}
\description{
Computes standardized (z-scored) values, handling NAs gracefully.
}
\examples{
zscore(c(1, 2, 3, 4, 5))
zscore(c(1, NA, 3))

}
